Graph identification cleanup
